Refactor shared utilities into dedicated modules
